+++ src/lib/assets/yaml/generateAllYaml.ts
