+++ src/lib/assets/yaml/generateZarfYamlChart.ts
